refactor g123 functions to separate class
